pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = "Simple MCP workshop demo with weather checking"
requires-python = ">=3.13"

dependencies = [
    "fastmcp>=2.0.0",
    "requests>=2.31.0", 
    "python-dotenv>=1.0.0",
]

# This is a script-only project, no packages to build
[tool.hatch.build]
exclude = ["*"]